# Reduce allocations, improve readEntries performance by simplifying loop