--- a/tsdb/engine/tsm1/reader.go
+++ b/tsdb/engine/tsm1/reader.go
@@ -61,14 +61,14 @@ type TSMIndex interface {
 	Entries(key []byte) []IndexEntry
 
 	// ReadEntries reads the index entries for key into entries.
-	ReadEntries(key []byte, entries *[]IndexEntry)
+	ReadEntries(key []byte, entries *[]IndexEntry) []IndexEntry
 
 	// Entry returns the index entry for the specified key and timestamp.  If no entry
 	// matches the key and timestamp, nil is returned.
 	Entry(key []byte, timestamp int64) *IndexEntry
 
-	// Key returns the key in the index at the given position.
-	Key(index int) ([]byte, byte, []IndexEntry)
+	// Key returns the key in the index at the given position, using entries to avoid allocations.
+	Key(index int, entries *[]IndexEntry) ([]byte, byte, []IndexEntry)
 
 	// KeyAt returns the key in the index at the given position.
 	KeyAt(index int) ([]byte, byte)
@@ -119,6 +119,7 @@ type BlockIterator struct {
 	n int
 
 	key     []byte
+	cache   []IndexEntry
 	entries []IndexEntry
 	err     error
 	typ     byte
@@ -149,7 +150,7 @@ func (b *BlockIterator) Next() bool {
 	}
 
 	if b.n-b.i > 0 {
-		b.key, b.typ, b.entries = b.r.Key(b.i)
+		b.key, b.typ, b.entries = b.r.Key(b.i, &b.cache)
 		b.i++
 
 		if len(b.entries) > 0 {
@@ -265,8 +266,8 @@ func (t *TSMReader) Path() string {
 }
 
 // Key returns the key and the underlying entry at the numeric index.
-func (t *TSMReader) Key(index int) ([]byte, byte, []IndexEntry) {
-	return t.index.Key(index)
+func (t *TSMReader) Key(index int, entries *[]IndexEntry) ([]byte, byte, []IndexEntry) {
+	return t.index.Key(index, entries)
 }
 
 // KeyAt returns the key and key type at position idx in the index.
@@ -487,8 +488,8 @@ func (t *TSMReader) Entries(key []byte) []IndexEntry {
 }
 
 // ReadEntries reads the index entries for key into entries.
-func (t *TSMReader) ReadEntries(key []byte, entries *[]IndexEntry) {
-	t.index.ReadEntries(key, entries)
+func (t *TSMReader) ReadEntries(key []byte, entries *[]IndexEntry) []IndexEntry {
+	return t.index.ReadEntries(key, entries)
 }
 
 // IndexSize returns the size of the index in bytes.
@@ -678,6 +679,11 @@ func (d *indirectIndex) search(key []byte) int {
 
 // Entries returns all index entries for a key.
 func (d *indirectIndex) Entries(key []byte) []IndexEntry {
+	return d.ReadEntries(key, nil)
+}
+
+// ReadEntries returns all index entries for a key.
+func (d *indirectIndex) ReadEntries(key []byte, entries *[]IndexEntry) []IndexEntry {
 	d.mu.RLock()
 	defer d.mu.RUnlock()
 
@@ -697,20 +703,21 @@ func (d *indirectIndex) Entries(key []byte) []IndexEntry {
 
 		// Read and return all the entries
 		ofs += n
-		var entries indexEntries
-		if _, err := readEntries(d.b[ofs:], &entries); err != nil {
+		var ie indexEntries
+		if entries != nil {
+			ie.entries = *entries
+		}
+		if _, err = readEntries(d.b[ofs:], &ie); err != nil {
 			panic(fmt.Sprintf("error reading entries: %v", err))
 		}
-		return entries.entries
+		if entries != nil {
+			*entries = ie.entries
+		}
+		return ie.entries
 	}
 
 	// The key is not in the index.  i is the index where it would be inserted.
 	return nil
-}
-
-// ReadEntries returns all index entries for a key.
-func (d *indirectIndex) ReadEntries(key []byte, entries *[]IndexEntry) {
-	*entries = d.Entries(key)
 }
 
 // Entry returns the index entry for the specified key and timestamp.  If no entry
@@ -726,7 +733,7 @@ func (d *indirectIndex) Entry(key []byte, timestamp int64) *IndexEntry {
 }
 
 // Key returns the key in the index at the given position.
-func (d *indirectIndex) Key(idx int) ([]byte, byte, []IndexEntry) {
+func (d *indirectIndex) Key(idx int, entries *[]IndexEntry) ([]byte, byte, []IndexEntry) {
 	d.mu.RLock()
 	defer d.mu.RUnlock()
 
@@ -741,11 +748,18 @@ func (d *indirectIndex) Key(idx int) ([]byte, byte, []IndexEntry) {
 
 	typ := d.b[int(ofs)+n]
 
-	var entries indexEntries
-	if _, err := readEntries(d.b[int(ofs)+n:], &entries); err != nil {
+	var ie indexEntries
+	if entries != nil {
+		ie.entries = *entries
+	}
+	if _, err = readEntries(d.b[int(ofs)+n:], &ie); err != nil {
 		return nil, 0, nil
 	}
-	return key, typ, entries.entries
+	if entries != nil {
+		*entries = ie.entries
+	}
+
+	return key, typ, ie.entries
 }
 
 // KeyAt returns the key in the index at the given position.
@@ -1451,19 +1465,21 @@ func readEntries(b []byte, entries *indexEntries) (n int, err error) {
 	count := int(binary.BigEndian.Uint16(b[n : n+indexCountSize]))
 	n += indexCountSize
 
-	entries.entries = make([]IndexEntry, count)
-	for i := 0; i < count; i++ {
-		var ie IndexEntry
-		start := i*indexEntrySize + indexCountSize + indexTypeSize
-		end := start + indexEntrySize
-		if end > len(b) {
-			return 0, fmt.Errorf("readEntries: data too short for indexEntry %d", i)
-		}
-		if err := ie.UnmarshalBinary(b[start:end]); err != nil {
+	if cap(entries.entries) < count {
+		entries.entries = make([]IndexEntry, count)
+	} else {
+		entries.entries = entries.entries[:count]
+	}
+
+	b = b[indexCountSize+indexTypeSize:]
+	for i := 0; i < len(entries.entries); i++ {
+		if err = entries.entries[i].UnmarshalBinary(b); err != nil {
 			return 0, fmt.Errorf("readEntries: unmarshal error: %v", err)
 		}
-		entries.entries[i] = ie
-		n += indexEntrySize
+		b = b[indexEntrySize:]
 	}
+
+	n += count * indexEntrySize
+
 	return
 }