Reduce allocations, improve `readEntries` performance by simplifying loop

* callers of `ReadEntries` and `Key` API can cache allocated slice
* `BlockIterator#Next` allocations should be significantly reduced in real-world usage too

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkIndirectIndex_Entries-8         13028         10439         -19.87%
BenchmarkIndirectIndex_ReadEntries-8     12905         7192          -44.27%
BenchmarkBlockIterator_Next-8            16374000      10879341      -33.56%

benchmark                                old allocs     new allocs     delta
BenchmarkIndirectIndex_Entries-8         1              1              +0.00%
BenchmarkIndirectIndex_ReadEntries-8     1              0              -100.00%
BenchmarkBlockIterator_Next-8            1001           2              -99.80%

benchmark                                old bytes     new bytes     delta
BenchmarkIndirectIndex_Entries-8         32768         32768         +0.00%
BenchmarkIndirectIndex_ReadEntries-8     32768         0             -100.00%
BenchmarkBlockIterator_Next-8            32768105      32896         -99.90%
```

###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated